[ModelExecution*] move classes into model_execution.py
